refactor(cli): extract resource-policy classification and messages from god file (#9279)
`resource_policy.rs` exceeded the 1500-line god-file threshold. Split its
two most cohesive, decision-free groups into focused submodules under
`resource_policy/`:

- `messages.rs`: human-facing warning/remediation builders (`severity_str`,
  `append_local_placement`, `primary_action`, `warning_message`, and the
  private `primary_reason`) — pure formatting over already-decided inputs.
- `classification.rs`: command-classification predicates
  (`is_controller_owned_fanout_coordination`,
  `is_lab_offloadable_fanout_coordinator`, `is_plan_only_command`,
  `is_read_only_agent_task`, `is_local_registry_management`) — pure matches
  over the parsed command tree that `hot_command` composes.

`resource_policy.rs` becomes `resource_policy/mod.rs` (1443 lines, under
threshold with margin). The split is behavior-preserving; all 42
resource_policy tests pass unchanged.
